requirements.txt
# DroneDeploy GTM Exercise Dependencies

# Core dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas==2.1.4
tqdm==4.66.1

# API clients
openai==1.99.1
anthropic==0.61.0
tavily-python==0.3.0

# Utilities
requests==2.31.0
lxml==4.9.3